# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"tags":[{"name":"Usuarios","description":"Operaciones relacionadas con los usuarios"}],"paths":{"/autonomaX/usuario/{id}":{"get":{"tags":["Usuarios"],"summary":"Obtener un usuario","description":"Dado un id, devuelve un el usuario, en caso contrario no devuelve nada","operationId":"obtenerUsuarioPorID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuarios"}}}}}},"put":{"tags":["Usuarios"],"summary":"Actualizar un usuario","description":"Actualiza al usuario mediante un id y un objeto Usuario (obligatorio esto para actualizar)","operationId":"actualizarUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuarios"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Usuarios"],"summary":"Eliminar un usuario","description":"Elimina al usuario de la base de datos, si no lo encuentra no hace nada","operationId":"BorrarUsuarios","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/autonomaX/nuevo_usuario":{"post":{"tags":["Usuarios"],"summary":"Registrar un nuevo usuario","description":"Crea un nuevo usuario validando el email, el nombre y la contraseña. Devuelve el usuario creado si todo es correcto.","operationId":"crearNuevoUsuario","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuarios"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/autonomaX/login":{"post":{"tags":["Usuarios"],"summary":"Loguear un usuario","description":"A partir de un email y una contraseña devuelve una respuesta afirmativa o negativa","operationId":"Login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuarios"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/autonomaX/usuarios":{"get":{"tags":["Usuarios"],"summary":"Devolver todos los usuarios","description":"Devuelve todos los usuarios","operationId":"ObtenerTodosLosUsuarios","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Usuarios"}}}}}}}}},"components":{"schemas":{"Usuarios":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre_usuario":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"rol":{"type":"string","enum":["Administrador","Usuario","Desarollo"]}}}}}}